moved get_pipeline_config_for_env_name_and_config_parser into kubernetes_pipeline_config.py
