+++ dags/kubernetes_pipeline.py
@@ -7,27 +7,15 @@ from typing import Sequence
 import airflow
 from airflow.providers.cncf.kubernetes.operators.pod import KubernetesPodOperator
 
-from data_pipeline.kubernetes.kubernetes_pipeline_config import MultiKubernetesPipelineConfig
+from data_pipeline.kubernetes.kubernetes_pipeline_config import (
+    get_multi_kubernetes_pipeline_config
+)
 from data_pipeline.utils.dags.data_pipeline_dag_utils import (
     create_dag,
 )
-from data_pipeline.utils.pipeline_config import (
-    get_deployment_env,
-    get_pipeline_config_for_env_name_and_config_parser
-)
+from data_pipeline.utils.pipeline_config import get_deployment_env
 
 LOGGER = logging.getLogger(__name__)
-
-
-class KubernetesPipelineConfigEnvironmentVariables:
-    CONFIG_FILE_PATH = 'KUBERNETES_PIPELINE_CONFIG_FILE_PATH'
-
-
-def get_multi_kubernetes_pipeline_config() -> MultiKubernetesPipelineConfig:
-    return get_pipeline_config_for_env_name_and_config_parser(
-        KubernetesPipelineConfigEnvironmentVariables.CONFIG_FILE_PATH,
-        MultiKubernetesPipelineConfig.from_dict
-    )
 
 
 def create_kubernetes_pipeline_dags() -> Sequence[airflow.DAG]:
